MS-03: New entity added; concert, migrations grouping, repository pattern with generics, eager loading

+++ src/MusicStore.API/Controllers/GenresController.cs
@@ -29,7 +29,15 @@ namespace MusicStore.API.Controllers
 
             try
             {
-                response.Data = await repository.GetAsync();
+                //mapping
+                var genresDb = await repository.GetAsync();
+                var genres = genresDb.Select(g => new GenreResponseDto
+                {
+                    Id = g.Id,
+                    Name = g.Name,
+                    Status = g.Status
+                }).ToList();
+                response.Data = genres;
                 response.Success = true;
                 logger.LogInformation("Obteniendo todos los generos musicales");
                 return Ok(response);
@@ -48,10 +56,23 @@ namespace MusicStore.API.Controllers
             var response = new BaseResponseGeneric<GenreResponseDto>();
             try
             {
-                response.Data = await repository.GetAsync(id);
+                var genresDb = await repository.GetAsync(id);
+                if(genresDb is null)
+                {
+                    logger.LogWarning($"No se encontró el genero musical con id {id}");
+                    response.ErrorMessage = "No se encontró el registro";
+                    return NotFound(response);
+                }
+                var genre = new GenreResponseDto
+                {
+                    Id = genresDb.Id,
+                    Name = genresDb.Name,
+                    Status = genresDb.Status
+                };
+                response.Data = genre;
                 response.Success = true;
                 logger.LogInformation($"Obteniendo el genero musical con id {id}");
-                return response.Data == null ? NotFound() : Ok(response);
+                return response.Data == null ? NotFound(response) : Ok(response);
             }
             catch (Exception ex)
             {
@@ -62,12 +83,17 @@ namespace MusicStore.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(GenreRequestDto genre)
+        public async Task<IActionResult> Post(GenreRequestDto genreRequestDto)
         {
             var response = new BaseResponseGeneric<int>();
             try
             {
-                var genreId = await repository.AddAsync(genre);
+                var genreDb =  new Genre()
+                {
+                    Name = genreRequestDto.Name,
+                    Status = genreRequestDto.Status
+                };
+                var genreId = await repository.AddAsync(genreDb);
                 response.Data = genreId;
                 response.Success = true;
                 logger.LogInformation($"Género músical insertado con id {genreId}");
@@ -81,12 +107,21 @@ namespace MusicStore.API.Controllers
             }
         }
         [HttpPut("{id:int}")]
-        public async Task<IActionResult> Put(int id, GenreRequestDto genre)
+        public async Task<IActionResult> Put(int id, GenreRequestDto genreRequestDto)
         {
             var response = new BaseResponse();
             try
             {
-                await repository.UpdateAsync(id, genre);
+                var genreDb = await repository.GetAsync(id);
+                if (genreDb is null)
+                {
+                    logger.LogWarning($"No se encontró el genero musical con id {id}");
+                    response.ErrorMessage = "No se encontró el registro";
+                    return NotFound(response);
+                }
+                genreDb.Name = genreRequestDto.Name;
+                genreDb.Status = genreRequestDto.Status;
+                await repository.UpdateAsync();
                 response.Success = true;
                 logger.LogInformation($"Género músical actualizado con id {id}");
                 return Ok(response);
@@ -104,6 +139,13 @@ namespace MusicStore.API.Controllers
             var response = new BaseResponse();
             try
             {
+                var genreDb = await repository.GetAsync(id);
+                if (genreDb is null)
+                {
+                    logger.LogWarning($"No se encontró el genero musical con id {id}");
+                    response.ErrorMessage = "No se encontró el registro";
+                    return NotFound(response);
+                }
                 await repository.DeleteAsync(id);
                 response.Success = true;
                 logger.LogInformation($"Género músical eliminado con id {id}");
